backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "complipilot",email = "founder@complipilot.co"}
]
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "pydantic[dotenv] (>=2.11.3,<3.0.0)",
    "openai (>=1.76.0,<2.0.0)",
    "alembic (>=1.15.2,<2.0.0)",
    "sqlmodel (>=0.0.24,<0.0.25)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)"
]

[tool.poetry]
packages = [{include = "backend", from = "src"}]


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
